Implement email verification check and User resource updates
The User model has been updated with a new 'verified' method to verify email status. Changes were also made to the UserFactory that allows the creation of users with a certain chance of having a verified email. Introduced updates to the UserResource which includes adding fields to the table and filter display. The Super Admin model has been updated in the DatabaseSeeder.