Add PictureResource related pages and configuration
A set of new pages has been added under the 'PictureResource' namespace, including CreatePicture, EditPicture, and ListPictures. In addition, the configuration for the PictureResource has been established, setting up the form, table, and relations for this resource. This is an important step in expanding the functionality of the picture management system.