Update Picture model to retrieve image URL
The Picture model has been updated by replacing a placeholder comment with an actual command to retrieve the image URL using Laravel's Storage facade. As part of the update, a new 'images' disk configuration was added to the filesystems configuration to specify the storage path and URL for images. This change enables the application to actively