Refactor database seeding and migration setup
This commit introduces multiple changes including the usage of Album factory in DatabaseSeeder, changing the superadmin user's email, and addition of 'view_count' in pictures table. Additionally, improvements were made to handling unauthenticated user id scenario in activities table and alterations in PicturesRelationManager to create a table with one column.